Cloned the Repo
the code base/project for the assignment can be found on GitHub, so you should start by cloning the hw3_s22 branch in that repository so you can work on it locally. You'll then want to create your own repository where you'll make all commits as you work on the assignment.